Rename params pkg to parameters
It was conflicting with the best variable name. 😆
