+++ pkg/parameters/parameters_test.go
@@ -1,4 +1,4 @@
-package params
+package parameters
 
 import (
 	"reflect"
